Refactor so that 'defer_function' is simply a StencilBuilder option
