+++ src/gt4py/stencil_builder.py
@@ -13,7 +13,6 @@ from gtc.passes.gtir_pipeline import GtirPipeline
 if TYPE_CHECKING:
     from gt4py.backend.base import Backend as BackendType
     from gt4py.backend.base import CLIBackendMixin
-    from gt4py.caching import CachingStrategy
     from gt4py.frontend.base import Frontend as FrontendType
     from gt4py.ir import StencilDefinition, StencilImplementation
     from gt4py.stencil_object import StencilObject
@@ -57,19 +56,24 @@ class StencilBuilder:
         self.options = options or BuildOptions(  # type: ignore
             **self.default_options_dict(definition_func)
         )
+        self._defer_function = (
+            options.backend_opts.pop("defer_function")
+            if options and "defer_function" in options.backend_opts
+            else None
+        )
         self.backend: "BackendType" = (
             backend(self) if backend else gt4py.backend.from_name("debug")(self)
         )
         self.frontend: "FrontendType" = frontend or gt4py.frontend.from_name("gtscript")
-        self.caching = self.get_caching_strategy(options)
+        self.caching = gt4py.caching.strategy_factory("jit", self)
         self._build_data: Dict[str, Any] = {}
         self._externals: Dict[str, Any] = {}
 
     def build(self) -> Type["StencilObject"]:
         """Generate, compile and/or load everything necessary to provide a usable stencil class."""
         # load, defer, or generate
-        if self.caching.is_deferred():
-            stencil_class = self.caching.defer()
+        if self._defer_function:
+            stencil_class = self._defer_function(self)
         else:
             stencil_class = None if self.options.rebuild else self.backend.load()
         if stencil_class is None:
@@ -83,14 +87,6 @@ class StencilBuilder:
     def generate_bindings(self, targe_language: str) -> Dict[str, Union[str, Dict]]:
         """Generate ``target_language`` bindings source, fail if backend does not support CLI."""
         return self.cli_backend.generate_bindings(targe_language)
-
-    def get_caching_strategy(self, options: BuildOptions) -> Type["CachingStrategy"]:
-        strategy_name = "jit"
-        strategy_opts = {}
-        if options and "defer_function" in options.backend_opts:
-            strategy_name = "deferred"
-            strategy_opts["defer_function"] = options.backend_opts.pop("defer_function")
-        return gt4py.caching.strategy_factory(strategy_name, self, **strategy_opts)
 
     def with_caching(
         self: "StencilBuilder", caching_strategy_name: str, *args: Any, **kwargs: Any
@@ -137,6 +133,7 @@ class StencilBuilder:
         """
         self._build_data = {}
         # mypy attribkwclass bug
+        self._defer_function = kwargs.pop("defer_function", None)
         self.options = BuildOptions(name=name, module=module, **kwargs)  # type: ignore
         return self
